requirements.txt
# Web Scraping
scrapy>=2.11.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
pymongo>=4.6.0

# Object Storage
minio>=7.2.0

# Orchestration
dagster>=1.6.0
dagster-webserver>=1.6.0

# Configuration
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0

# Date handling
python-dateutil>=2.8.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Type checking
mypy>=1.8.0

# Code formatting
black>=24.0.0
isort>=5.13.0
